+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the library will be documented in this file.
 
+## vX.X.X (Month DD, YYYY)
+
+- Change output of empty string dates to `null` (issue #3)
+
 ## v0.2.0 (October 11, 2023)
 
 - Add support for `false` booleans to `getFieldValue` util (issue #1)